Clean up artifacts
The BoF site only uses the `Conference program page` entries.  The others are for the IDEAS and CASS websites where they appear as events.

Remove the link to the BoF series on the assumption that if someone is interested enough, they'll go to the BoF webpage and they can figure out the series from there.

Reorder the artifact entries so that Conference program page comes last and the blog article comes first.

Tighten up the label for the BoF webpage and eliminate the need for the format entry.